refactor(pocha-manage): colocate cross-field validation trigger in PochaInfoFields (P4)
